+++ tests/testthat/helper-httptest.R
@@ -5,8 +5,8 @@ library(httptest)
 # - cache: Use cached requests in `testthat/tests/api.tiingo.com`
 # - capture: Run the actual tests, capturing as you go
 # - full: Run the actual tests, without any capturing
-local_riingo_mock <- function(envir = parent.frame()) {
-  type <- Sys.getenv("RIINGO_MOCK_TYPE", unset = "cache")
+local_riingo_mock <- function(type = NULL, envir = parent.frame()) {
+  type <- type %||% Sys.getenv("RIINGO_MOCK_TYPE", unset = "cache")
 
   local_mocker <- switch(type,
     cache = local_riingo_mock_cache,